+++ data-browser/src/components/Button.tsx
@@ -140,10 +140,11 @@ export const ButtonMargin = styled(ButtonBase) <ButtonProps>`
   &:focus-visible:not([disabled]),
   &:hover:not([disabled]) {
     box-shadow: ${p => p.theme.boxShadowIntense};
-    background-color: ${p => (p.subtle ? 'initial' : p.theme.colors.mainLight)};
+    background-color: ${p =>
+    p.subtle ? p.theme.colors.bg : p.theme.colors.mainLight};
     color: ${p => (p.subtle ? p.theme.colors.main : p.theme.colors.bg)};
     border-color: ${p =>
-      p.subtle ? p.theme.colors.main : p.theme.colors.mainLight};
+    p.subtle ? p.theme.colors.main : p.theme.colors.mainLight};
   }
 
   &:active:not([disabled]) {